requirements.txt
wxPython
igraph
numpy==1.25.2
opencv-python==4.11.0.86
rosbags
tqdm
pyyaml
pycairo
scipy